requirements.txt
Django
whitenoise
django-storages[azure]
asgiref
attrs
Automat
cachetools
certifi
cffi
channels
chardet
constantly
coreapi
coreschema
cryptography
daphne
djangochannelsrestframework
djangorestframework
djangorestframework-simplejwt
pyJWT
drf-yasg
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
httplib2
hyperlink
idna
incremental
inflection
itypes
Jinja2
MarkupSafe
packaging
Pillow
protobuf
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
pyOpenSSL
pyparsing
python-dotenv
pytz
requests
rsa
ruamel.yaml
ruamel.yaml.clib
service-identity
six
sqlparse
Twisted
txaio
openpyxl
mysqlclient
gunicorn
dj-database-url
